Update Dictionarymethods.rst
Added the sentence, "However, it is important to note that calling these dictionary method requires the inclusion of a pair of parenthesis- ''()'' at the end of the call.  The format would look something like this:".  I personally thought this was confusing the first time I saw this in the book.  I think it would be better if we explained the syntax up front, and maybe the reasoning behind it (I am not sure of what the reasoning exactly is), instead of leaving it for them to pick up on it in the code.